requirements.txt
streamlit
PyMuPDF
langchain
google-generativeai
langchain-google-genai
langchain-prompts
langchain-chains